Finished frontend design
